+++ plugins/PayPal/src/PayPalPlugin.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace LifeWheel\Plugins\PayPal;
+
+use App\Plugins\BasePlugin;
+use App\Plugins\PluginContext;
+
+final class PayPalPlugin extends BasePlugin
+{
+    public function install(PluginContext $context): void
+    {
+        //
+    }
+
+    public function activate(PluginContext $context): void
+    {
+        //
+    }
+
+    public function deactivate(PluginContext $context): void
+    {
+        //
+    }
+}